Huge patch:   * Compile with Wall, Wextra, Werror      - Remove unused variables      - Enable large file support      - Fix esps read/write (inside a structure there can be padding bytes that were being ignored)

+++ base_class/EST_DMatrix.cc
@@ -68,7 +68,7 @@ EST_DMatrix::EST_DMatrix(const EST_DMatrix &a, int b)
 
 EST_DMatrix & EST_DMatrix::operator+=(const EST_DMatrix &a)
 {
-    int i, j;
+    ssize_t i, j;
     if (a.num_columns() != num_columns())
     {
 	cerr <<"Matrix addition error: bad number of columns\n";
@@ -88,7 +88,7 @@ EST_DMatrix & EST_DMatrix::operator+=(const EST_DMatrix &a)
 
 EST_DMatrix & EST_DMatrix::operator-=(const EST_DMatrix &a)
 {
-    int i, j;
+    ssize_t i, j;
     if (a.num_columns() != num_columns())
     {
 	cerr <<"Matrix subtraction error: bad number of columns\n";
@@ -109,7 +109,7 @@ EST_DMatrix & EST_DMatrix::operator-=(const EST_DMatrix &a)
 EST_DMatrix & EST_DMatrix::operator*=(const double f)
 {
 
-    int i,j;
+    ssize_t i,j;
     for (i = 0; i < num_rows(); ++i)
 	for (j = 0; j < num_columns(); ++j)
 	    a_no_check(i, j) *= f;
@@ -120,7 +120,7 @@ EST_DMatrix & EST_DMatrix::operator*=(const double f)
 EST_DMatrix & EST_DMatrix::operator/=(const double f)
 {
 
-    int i,j;
+    ssize_t i,j;
     for (i = 0; i < num_rows(); ++i)
 	for (j = 0; j < num_columns(); ++j)
 	    a_no_check(i, j) /= f;
@@ -131,7 +131,7 @@ EST_DMatrix & EST_DMatrix::operator/=(const double f)
 EST_DMatrix operator+(const EST_DMatrix &a, const EST_DMatrix &b)
 {
     EST_DMatrix ab;
-    int i, j;
+    ssize_t i, j;
     if (a.num_columns() != b.num_columns())
     {
 	cerr <<"Matrix addition error: bad number of columns\n";
@@ -153,7 +153,7 @@ EST_DMatrix operator+(const EST_DMatrix &a, const EST_DMatrix &b)
 EST_DMatrix operator-(const EST_DMatrix &a,const EST_DMatrix &b)
 {
     EST_DMatrix ab;
-    int i, j;
+    ssize_t i, j;
 
     if (a.num_columns() != b.num_columns())
     {
@@ -182,7 +182,7 @@ EST_DMatrix operator*(const double x, const EST_DMatrix &a)
 EST_DMatrix operator*(const EST_DMatrix &a, const double x)
 {
     EST_DMatrix b(a, 0);
-    int i, j;
+    ssize_t i, j;
 
     for (i = 0; i < a.num_rows(); ++i)
 	for (j = 0; j < a.num_columns(); ++j)
@@ -206,7 +206,7 @@ EST_DVector operator*(const EST_DMatrix &a, const EST_DVector &v)
 	return b;
     }
 
-    int i, j;
+    ssize_t i, j;
     for (i = 0; i < a.num_rows(); ++i){
 	b[i] = 0.0;
 	for (j = 0; j < a.num_columns(); ++j)
@@ -218,7 +218,7 @@ EST_DVector operator*(const EST_DMatrix &a, const EST_DVector &v)
 EST_DVector operator+(const EST_DVector &a, const EST_DVector &b)
 {
     EST_DVector ab;
-    int i;
+    ssize_t i;
     if (a.length() != b.length())
     {
 	cerr <<"Vector addition error: mismatched lengths\n";
@@ -235,7 +235,7 @@ EST_DVector operator+(const EST_DVector &a, const EST_DVector &b)
 EST_DVector operator-(const EST_DVector &a, const EST_DVector &b)
 {
     EST_DVector ab;
-    int i;
+    ssize_t i;
     if (a.length() != b.length())
     {
 	cerr <<"Vector subtraction error: mismatched lengths\n";
@@ -265,7 +265,7 @@ EST_DVector operator*(const EST_DVector &v,const EST_DMatrix &a)
 	return b;
     }
 
-    int i, j;
+    ssize_t i, j;
     for (j = 0; j < a.num_columns(); ++j){
 	b[j] = 0.0;
 	for (i = 0; i < a.num_rows(); ++i)
@@ -299,7 +299,7 @@ void multiply(const EST_DMatrix &a, const EST_DMatrix &b, EST_DMatrix &ab)
     }
 
     ab.resize(a.num_rows(), b.num_columns());
-    int i, j, k, n;
+    ssize_t i, j, k, n;
     n = a.num_columns();	// could also be b.num_rows()
     
     for (i = 0; i < a.num_rows(); ++i)
@@ -312,9 +312,9 @@ void multiply(const EST_DMatrix &a, const EST_DMatrix &b, EST_DMatrix &ab)
 	}
 }
 
-void EST_DMatrix::copyin(double **inx, int rows, int cols)
+void EST_DMatrix::copyin(double **inx, ssize_t rows, ssize_t cols)
 {
-    int i, j;
+    ssize_t i, j;
 
     resize(rows, cols);
 
@@ -332,7 +332,7 @@ EST_write_status EST_DMatrix::save(const EST_String &filename,
 	return est_save(filename,type);
     else
     {   // the old stuff raw unheadered
-	int i, j;
+	ssize_t i, j;
 	ostream *outf;
 	if (filename == "-")
 	    outf = &cout;
@@ -365,7 +365,7 @@ EST_write_status EST_DMatrix::est_save(const EST_String &filename,
 				       const EST_String &type)
 {
     // Binary save with short header for byte swap and sizes
-    int i,j;
+    ssize_t i,j;
     FILE *fd;
     if (filename == "-")
 	fd = stdout;
@@ -389,8 +389,8 @@ EST_write_status EST_DMatrix::est_save(const EST_String &filename,
     else
 	fprintf(fd,"DataType ascii\n");
 
-    fprintf(fd,"rows %d\n",num_rows());
-    fprintf(fd,"columns %d\n",num_columns());
+    fprintf(fd,"rows %zu\n",num_rows());
+    fprintf(fd,"columns %zu\n",num_columns());
 
     fprintf(fd,"EST_Header_End\n");
 
@@ -426,7 +426,7 @@ EST_read_status EST_DMatrix::est_load(const EST_String &filename)
 {
 
   // ascii/binary load with short header for byte swap and sizes
-  int i,j,k;
+  ssize_t i,j,k;
   int rows, cols, swap;
   EST_TokenStream ts;
   EST_read_status r;
@@ -508,7 +508,7 @@ EST_read_status EST_DMatrix::load(const EST_String &filename)
     {   // maybe its an ancient ascii file
 	EST_TokenStream ts, tt;
 	EST_StrList sl;
-	int i, j, n_rows=0, n_cols=0;
+	ssize_t i, j, n_rows=0, n_cols=0;
 	EST_String t;
 	EST_Litem *p;
 	if (((filename == "-") ? ts.open(cin) : ts.open(filename)) != 0)
@@ -556,7 +556,7 @@ EST_read_status EST_DMatrix::load(const EST_String &filename)
 EST_read_status EST_DVector::est_load(const EST_String &filename)
 {    
   // ascii/binary load with short header for byte swap and sizes
-  int i,k;
+  ssize_t i,k;
   int l, swap;
   EST_TokenStream ts;
   EST_read_status r;
@@ -671,7 +671,7 @@ EST_read_status EST_DVector::load(const EST_String &filename)
 
 EST_DVector & EST_DVector::operator+=(const EST_DVector &s)
 {
-    int i;
+    ssize_t i;
     if(n() != s.n()){
 	cerr << "Cannot elementwise add vectors of differing lengths" 
 	    << endl;
@@ -694,7 +694,7 @@ EST_DVector& EST_DVector::operator*=(const EST_DVector &s)
 	return *this;
     }
 
-    for (int i = 0; i < n(); ++i)
+    for (ssize_t i = 0; i < n(); ++i)
 	(*this)[i] *= s(i);
 
     return *this;
@@ -702,7 +702,7 @@ EST_DVector& EST_DVector::operator*=(const EST_DVector &s)
 
 EST_DVector& EST_DVector::operator*=(const double f)
 {
-    for (int i = 0; i < n(); ++i)
+    for (ssize_t i = 0; i < n(); ++i)
 	(*this)[i] *= f;
 
     return *this;
@@ -716,7 +716,7 @@ double operator*(const EST_DVector &v1, const EST_DVector &v2)
       return 0;
     }
   double p=0;
-  for (int i = 0; i < v1.length(); ++i)
+  for (ssize_t i = 0; i < v1.length(); ++i)
       p += v1.a_no_check(i) * v2.a_no_check(i);
 
     return p;
@@ -725,7 +725,7 @@ double operator*(const EST_DVector &v1, const EST_DVector &v2)
 
 EST_DVector& EST_DVector::operator/=(const double f)
 {
-    for (int i = 0; i < n(); ++i)
+    for (ssize_t i = 0; i < n(); ++i)
 	(*this)[i] /= f;
 
     return *this;
@@ -740,7 +740,7 @@ EST_write_status EST_DVector::save(const EST_String &filename,
 	return est_save(filename,type);
     else
     {   // the old stuff raw unheadered
-	int i;
+	ssize_t i;
 	ostream *outf;
 	if (filename == "-")
 	    outf = &cout;
@@ -770,7 +770,7 @@ EST_write_status EST_DVector::est_save(const EST_String &filename,
 				      const EST_String &type)
 {
     // Binary save with short header for byte swap and sizes
-    int i;
+    ssize_t i;
     FILE *fd;
     if (filename == "-")
 	fd = stdout;
@@ -794,7 +794,7 @@ EST_write_status EST_DVector::est_save(const EST_String &filename,
     else
 	fprintf(fd,"DataType ascii\n");
 
-    fprintf(fd,"length %d\n",length());
+    fprintf(fd,"length %zu\n",length());
     fprintf(fd,"EST_Header_End\n");
 
     if (type == "est_binary")
