+++ base_class/EST_FMatrix.cc
@@ -73,7 +73,7 @@ EST_FMatrix::EST_FMatrix(const EST_FMatrix &a, int b)
 
 EST_FMatrix & EST_FMatrix::operator+=(const EST_FMatrix &a)
 {
-    int i, j;
+    ssize_t i, j;
     if (a.num_columns() != num_columns())
     {
 	cerr <<"Matrix addition error: bad number of columns\n";
@@ -93,7 +93,7 @@ EST_FMatrix & EST_FMatrix::operator+=(const EST_FMatrix &a)
 
 EST_FMatrix & EST_FMatrix::operator-=(const EST_FMatrix &a)
 {
-    int i, j;
+    ssize_t i, j;
     if (a.num_columns() != num_columns())
     {
 	cerr <<"Matrix subtraction error: bad number of columns\n";
@@ -114,7 +114,7 @@ EST_FMatrix & EST_FMatrix::operator-=(const EST_FMatrix &a)
 EST_FMatrix & EST_FMatrix::operator*=(const float f)
 {
 
-    int i,j;
+    ssize_t i,j;
     for (i = 0; i < num_rows(); ++i)
 	for (j = 0; j < num_columns(); ++j)
 	    a_no_check(i, j) *= f;
@@ -125,7 +125,7 @@ EST_FMatrix & EST_FMatrix::operator*=(const float f)
 EST_FMatrix & EST_FMatrix::operator/=(const float f)
 {
 
-    int i,j;
+    ssize_t i,j;
     for (i = 0; i < num_rows(); ++i)
 	for (j = 0; j < num_columns(); ++j)
 	    a_no_check(i, j) /= f;
@@ -136,7 +136,7 @@ EST_FMatrix & EST_FMatrix::operator/=(const float f)
 EST_FMatrix operator+(const EST_FMatrix &a, const EST_FMatrix &b)
 {
     EST_FMatrix ab;
-    int i, j;
+    ssize_t i, j;
     if (a.num_columns() != b.num_columns())
     {
 	cerr <<"Matrix addition error: bad number of columns\n";
@@ -158,7 +158,7 @@ EST_FMatrix operator+(const EST_FMatrix &a, const EST_FMatrix &b)
 EST_FMatrix operator-(const EST_FMatrix &a,const EST_FMatrix &b)
 {
     EST_FMatrix ab;
-    int i, j;
+    ssize_t i, j;
 
     if (a.num_columns() != b.num_columns())
     {
@@ -182,7 +182,7 @@ EST_FMatrix operator-(const EST_FMatrix &a,const EST_FMatrix &b)
 EST_FMatrix operator*(const EST_FMatrix &a, const float x)
 {
     EST_FMatrix b(a, 0);
-    int i, j;
+    ssize_t i, j;
 
     for (i = 0; i < a.num_rows(); ++i)
 	for (j = 0; j < a.num_columns(); ++j)
@@ -194,7 +194,7 @@ EST_FMatrix operator*(const EST_FMatrix &a, const float x)
 int operator !=(const EST_FVector &fv1, 
 		const EST_FVector &fv2)
 {
-    int i;
+    ssize_t i;
     if(fv1.length() != fv2.length())
 	return FALSE;
     for(i=0;i<fv1.length();i++)
@@ -219,7 +219,7 @@ EST_FVector operator*(const EST_FMatrix &a, const EST_FVector &v)
 	return b;
     }
 
-    int i, j;
+    ssize_t i, j;
     for (i = 0; i < a.num_rows(); ++i){
 	b[i] = 0.0;
 	for (j = 0; j < a.num_columns(); ++j)
@@ -231,7 +231,7 @@ EST_FVector operator*(const EST_FMatrix &a, const EST_FVector &v)
 EST_FVector operator+(const EST_FVector &a, const EST_FVector &b)
 {
     EST_FVector ab;
-    int i;
+    ssize_t i;
     if (a.length() != b.length())
     {
 	cerr <<"Vector addition error: mismatched lengths\n";
@@ -248,7 +248,7 @@ EST_FVector operator+(const EST_FVector &a, const EST_FVector &b)
 EST_FVector operator-(const EST_FVector &a, const EST_FVector &b)
 {
     EST_FVector ab;
-    int i;
+    ssize_t i;
     if (a.length() != b.length())
     {
 	cerr <<"Vector subtraction error: mismatched lengths\n";
@@ -278,7 +278,7 @@ EST_FVector operator*(const EST_FVector &v,const EST_FMatrix &a)
 	return b;
     }
 
-    int i, j;
+    ssize_t i, j;
     for (j = 0; j < a.num_columns(); ++j){
 	b[j] = 0.0;
 	for (i = 0; i < a.num_rows(); ++i)
@@ -312,7 +312,7 @@ void multiply(const EST_FMatrix &a, const EST_FMatrix &b, EST_FMatrix &ab)
     }
 
     ab.resize(a.num_rows(), b.num_columns());
-    int i, j, k, n;
+    ssize_t i, j, k, n;
     n = a.num_columns();	// could also be b.num_rows()
     
     for (i = 0; i < a.num_rows(); ++i)
@@ -325,9 +325,9 @@ void multiply(const EST_FMatrix &a, const EST_FMatrix &b, EST_FMatrix &ab)
 	}
 }
 
-void EST_FMatrix::copyin(float **inx, int rows, int cols)
+void EST_FMatrix::copyin(float **inx, ssize_t rows, ssize_t cols)
 {
-    int i, j;
+    ssize_t i, j;
 
     resize(rows, cols);
 
@@ -345,7 +345,7 @@ EST_write_status EST_FMatrix::save(const EST_String &filename,
 	return est_save(filename,type);
     else
     {   // the old stuff raw unheadered
-	int i, j;
+	ssize_t i, j;
 	ostream *outf;
 	if (filename == "-")
 	    outf = &cout;
@@ -377,7 +377,7 @@ EST_write_status EST_FMatrix::est_save(const EST_String &filename,
 				       const EST_String &type)
 {
     // Binary save with short header for byte swap and sizes
-    int i,j;
+    ssize_t i,j;
     FILE *fd;
     if (filename == "-")
 	fd = stdout;
@@ -401,8 +401,8 @@ EST_write_status EST_FMatrix::est_save(const EST_String &filename,
     else
 	fprintf(fd,"DataType ascii\n");
 
-    fprintf(fd,"rows %d\n",num_rows());
-    fprintf(fd,"columns %d\n",num_columns());
+    fprintf(fd,"rows %zu\n",num_rows());
+    fprintf(fd,"columns %zu\n",num_columns());
 
     fprintf(fd,"EST_Header_End\n");
 
@@ -437,7 +437,7 @@ EST_write_status EST_FMatrix::est_save(const EST_String &filename,
 EST_read_status EST_FMatrix::est_load(const EST_String &filename)
 {
     // ascii/binary load with short header for byte swap and sizes
-    int i,j,k;
+    ssize_t i,j,k;
     int rows, cols, swap;
     EST_TokenStream ts;
     EST_read_status r;
@@ -520,7 +520,7 @@ EST_read_status EST_FMatrix::load(const EST_String &filename)
     {   // maybe its an ancient ascii file
 	EST_TokenStream ts, tt;
 	EST_StrList sl;
-	int i, j, n_rows=0, n_cols=0;
+	ssize_t i, j, n_rows=0, n_cols=0;
 	EST_String t;
 	EST_Litem *p;
 	if (((filename == "-") ? ts.open(cin) : ts.open(filename)) != 0)
@@ -566,7 +566,7 @@ EST_read_status EST_FMatrix::load(const EST_String &filename)
 
 EST_FVector & EST_FVector::operator+=(const EST_FVector &s)
 {
-    int i;
+    ssize_t i;
     if(n() != s.n()){
 	cerr << "Cannot elementwise add vectors of differing lengths" 
 	    << endl;
@@ -589,7 +589,7 @@ EST_FVector& EST_FVector::operator*=(const EST_FVector &s)
 	return *this;
     }
 
-    for (int i = 0; i < n(); ++i)
+    for (ssize_t i = 0; i < n(); ++i)
 	(*this)[i] *= s(i);
 
     return *this;
@@ -597,7 +597,7 @@ EST_FVector& EST_FVector::operator*=(const EST_FVector &s)
 
 EST_FVector& EST_FVector::operator*=(const float f)
 {
-    for (int i = 0; i < n(); ++i)
+    for (ssize_t i = 0; i < n(); ++i)
 	(*this)[i] *= f;
 
     return *this;
@@ -606,7 +606,7 @@ EST_FVector& EST_FVector::operator*=(const float f)
 
 EST_FVector& EST_FVector::operator/=(const float f)
 {
-    for (int i = 0; i < n(); ++i)
+    for (ssize_t i = 0; i < n(); ++i)
 	(*this)[i] /= f;
 
     return *this;
@@ -617,7 +617,7 @@ EST_FVector& EST_FVector::operator/=(const float f)
 EST_read_status EST_FVector::est_load(const EST_String &filename)
 {    
   // ascii/binary load with short header for byte swap and sizes
-  int i,k;
+  ssize_t i,k;
   int l, swap;
   EST_TokenStream ts;
   EST_read_status r;
@@ -822,7 +822,7 @@ float operator*(const EST_FVector &v1, const EST_FVector &v2)
 	return b;
     }
 
-    int i;
+    ssize_t i;
     for (i = 0; i < v1.length(); ++i)
 	b += v1.a_no_check(i) * v2.a_no_check(i);
 
@@ -838,7 +838,7 @@ EST_write_status EST_FVector::save(const EST_String &filename,
 	return est_save(filename,type);
     else
     {   // the old stuff raw unheadered
-	int i;
+	ssize_t i;
 	ostream *outf;
 	if (filename == "-")
 	    outf = &cout;
@@ -868,7 +868,7 @@ EST_write_status EST_FVector::est_save(const EST_String &filename,
 				      const EST_String &type)
 {
     // Binary save with short header for byte swap and sizes
-    int i;
+    ssize_t i;
     FILE *fd;
     if (filename == "-")
 	fd = stdout;
@@ -892,7 +892,7 @@ EST_write_status EST_FVector::est_save(const EST_String &filename,
     else
 	fprintf(fd,"DataType ascii\n");
 
-    fprintf(fd,"length %d\n",length());
+    fprintf(fd,"length %zu\n",length());
     fprintf(fd,"EST_Header_End\n");
 
     if (type == "est_binary")
