+++ base_class/EST_SMatrix.cc
@@ -68,7 +68,7 @@ int EST_SMatrix::rateconv(int in_samp_freq, int out_samp_freq)
   int *len = new int[num_columns()];
   int max_len=0;
 
-  for(int c=0; c<num_columns(); c++)
+  for(ssize_t c=0; c<num_columns(); c++)
     {
       short *out_buf;
       int osize;
@@ -92,7 +92,7 @@ int EST_SMatrix::rateconv(int in_samp_freq, int out_samp_freq)
   resize(max_len, EST_CURRENT, 0);
   fill(0);
 
-  for(int c1=0; c1<num_columns(); c1++)
+  for(ssize_t c1=0; c1<num_columns(); c1++)
     {
       set_column(c1, results[c1], 0, len[c1]);
       delete [] results[c1];
