+++ base_class/inst_tmpl/matrix_f_t.cc
@@ -54,7 +54,7 @@ Instantiate_TSimpleMatrix(float)
 
 EST_write_status save(const EST_String &filename, const EST_TMatrix<float> &a)
 {
-    int i,j;
+    ssize_t i,j;
     ostream *outf;
     EST_String s;
     if (filename == "-")
