+++ base_class/vec_mat_aux_i.cc
@@ -48,7 +48,7 @@ using namespace std;
 
 int matrix_max(const EST_IMatrix &a)
 {
-    int i, j;
+    ssize_t i, j;
     int v = INT_MIN;
     
     for (i = 0; i < a.num_rows(); ++i)
